refactor(**/*/downstream_fec.*): refact downstream_fec
