+++ app/controllers/forms/__init__.py
@@ -1,7 +1,6 @@
 from .network_form import NetworkForm
 from .auth_form import LoginForm, RegisterForm
 from .set_static_route_form import StaticRouteForm
-from .downstream_fec_form import Downstream_fec_form
 from .set_access_address_assignment import AddressAssignmentForm
 
 __all__ = [
@@ -9,6 +8,5 @@ __all__ = [
     'NetworkForm',
     'RegisterForm',
     'StaticRouteForm',
-    'Downstream_fec_form',
     'AddressAssignmentForm',
 ]
