--- app/controllers/netmiko/dmos/__init__.py
@@ -1,5 +1,0 @@
-from .no_downstream_fec import no_downstream_fec
-
-__all__ = [
-    'no_downstream_fec',
-]
