+++ app/controllers/routes/dmos/downstream_fec.py
@@ -4,9 +4,9 @@ from dotenv import load_dotenv
 from app.models import Users, Olts
 from cryptography.fernet import Fernet
 from flask import Blueprint, render_template, flash
-from app.controllers.forms import Downstream_fec_form
+from app.controllers.forms.downstream_fec import Downstream_fec_form
 from flask_login import current_user, login_required, fresh_login_required
-from app.controllers.netmiko.dmos.no_downstream_fec import no_downstream_fec
+from app.controllers.netmiko.dmos.downstream_fec import no_downstream_fec
 
 load_dotenv()
