+++ app/templates/vendors/dmos/downstream_fec.html
@@ -1,53 +1,54 @@
 {% extends "base.html" %}
 
 {% block content %}
-<h1 class="display-4">{% block title %}Downstream-fec{% endblock %}</h1>
+<h1>Downstream-fec</h1>
 
-{% block comment %}
-{{ super() }}
-{% endblock %}
 <hr>
 
 <form method="POST" action="#" autocomplete="on">
     {{ form.hidden_tag() }}
 
-    <!-- ROW OF THE HOSTNAME -->
-    <div class="form-row">
-        <div class="form-group col">{{ form.hostname(class="form-control") }}</div>
-        <div class="form-group col-.5">{{ form.chassis(class="form-control") }}</div>
-        <div class="form-group col-.5">{{ form.slot(class="form-control") }}</div>
-        <div class="form-group col-.5">{{ form.port_id(class="form-control") }}</div>
-        <div class="form-group col">{{ form.dmos_command(class="form-control") }}</div>
-    </div>
-
-
-    <div class="form-row">
-        <!-- Usuários -->
-        <div class="form-group col mb-3">
-            <input type="text" class="form-control" value="{{ current_user.username }}" readonly>
-        </div>
-
-        <!-- Passwords -->
-        <div class="input-group col mb-3">
-            <input type="password" class="form-control" id="password-input" value="{{ current_user.password[:12] }}"
-                readonly>
-            <div class="input-group-append">
-                <i class="fas fa-eye input-group-text" id="toggle-password"></i>
+    <div class="row mb-3">
+        <div class="col">
+            <div class="form-floating">
+                {{ form.hostname(class="form-control") }}
+                {{ form.hostname.label }}
             </div>
         </div>
     </div>
 
-    <div class="form-row">
-        <div class="form-group col">{{ form.submit(class="btn btn-primary") }}</div>
+    <div class="row g-3 mb-3">
+        <div class="col-2">
+            <div class="form-floating">
+                {{ form.chassis(class="form-control") }}
+                {{ form.chassis.label }}
+            </div>
+        </div>
+        <div class="col-2">
+            <div class="form-floating">
+                {{ form.slot(class="form-control") }}
+                {{ form.slot.label }}
+            </div>
+        </div>
+        <div class="col-2">
+            <div class="form-floating">
+                {{ form.port_id(class="form-control") }}
+                {{ form.port_id.label }}
+            </div>
+        </div>
+        <div class="col">
+            <div class="form-floating">
+                {{ form.dmos_command(class="form-control") }}
+                {{ form.dmos_command.label }}
+            </div>
+        </div>
     </div>
+
+    {{ form.submit(class="btn btn-primary") }}
 </form>
 
 <br>
 
-<pre class="text-center"> show interface gpon {chassis}/{slot}/{port_id} </pre>
-
-<pre class="max-height" style=" border: solid 1px rgba(0, 0, 0, .1); 
-                                    padding: 1rem; 
-                                    margin-top: 1rem; 
-                                    border-radius: 10px;">{{ output }}</pre>
+<smal class="text-muted"> show interface gpon {chassis}/{slot}/{port_id} </smal>
+<pre class="border border-3 rounded-3 p-3">{{ output }}</pre>
 {% endblock %}
